Deprecate `qmk generate-compilation-database`.
